dublin-bikes-analysis/requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
jupyter-client==5.3.1
jupyter-core==4.5.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
pandas==0.25.1
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
requests==2.22.0
scipy==1.3.1
six==1.12.0
statsmodels==0.10.1
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.5
wcwidth==0.1.7
